pyproject.toml
[project]
name = "blindnotdumb"
version = "0.1.0"
description = "Blind Not Dumb Hugo migration helpers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docutils>=0.22",
    "markdownify>=1.1.0",
    "pyyaml>=6.0.2",
]
